Incredible Leap Forward
Removed the last known limitation on school computers that use less than
2gb of ram. The game runs at 225MB now instead of 750MB, and has an
increased performance on gameplay yet again. The increase is anywhere
between 5 to 40 frames depending on CPU. This increase is rather huge in
smoother gameplay. All pictures have been resized to 50% image width and
height. So the game is less than 50 MB now to download. We can add
really cool features now since the game mechanics are pretty much set in
stone now. I am very happy with this version and hope to add a lot of
scary stuff soon. The old assets has been copied to my game computer and
will be finalized. Might be able to add the kicking animation soon and
whatever else we did not use.